refactor(efloat): generalize RoundingMode to sw/universal/behavior/rounding.hpp
Expose RoundingMode as a standard common enum class inside the
sw::universal namespace under Option A (<universal/behavior/rounding.hpp>).

This makes standard IEEE-754 rounding modes globally available to all number
systems in the Universal Library (e.g., lns, cfloat, and fixpnt), while
efloat_impl.hpp includes the new header cleanly and retains its thread_local
efloat_rounding_mode variable.

Relates to #1091, PR #1104
